check file or directory existance

+++ lib/restore.sh
@@ -51,21 +51,27 @@ then
     brew install neovim/neovim/neovim
 fi
 
-#install my vim conf
-curl https://raw.githubusercontent.com/Alexoner/spf13-vim/3.0/bootstrap.sh -L > spf13-vim.sh && sh spf13-vim.sh
-# install powerline fonts for vim
-git clone https://github.com/powerline/fonts.git /tmp/fonts && cd /tmp/fonts && sh /tmp/fonts/install.sh
-cd "$HOME/.synccf" || exit 0
+if [ ! -e "$HOME/.spf13-vim-3" ]; then
+    #install my vim conf
+    curl https://raw.githubusercontent.com/Alexoner/spf13-vim/3.0/bootstrap.sh -L > spf13-vim.sh && sh spf13-vim.sh
+    # install powerline fonts for vim
+    git clone https://github.com/powerline/fonts.git /tmp/fonts && cd /tmp/fonts && sh /tmp/fonts/install.sh
+    cd "$HOME/.synccf" || exit 0
+elif [ ! -d "$HOME/.spf13-vim-3" ]; then
+    echo "$HOME/.spf13-vim-3 exists but not a directory"
+fi
 
 #install shadowsocks-heroku
 SOFT_DIR="$HOME/Documents/soft"
-mkdir -p "$SOFT_DIR"
-git clone https://github.com/mrluanma/shadowsocks-heroku.git "$SOFT_DIR/shadowsocks-heroku"
-cd "$SOFT_DIR/shadowsocks-heroku" && npm install
-echo '#!/bin/sh' > run.sh
-echo 'node local.js -s xxx-island.herokuapp.com -l 1081 -m rc4 -k xxx11yy -r 80 1>> ss.log  2> ss.log' >> run.sh
-chmod +x run.sh
-cd "$HOME/.synccf" || exit 0
+if [ ! -e "$HOME/Documents/soft/shadowsocks-heroku" ]; then
+    mkdir -p "$SOFT_DIR"
+    git clone https://github.com/mrluanma/shadowsocks-heroku.git "$SOFT_DIR/shadowsocks-heroku"
+    cd "$SOFT_DIR/shadowsocks-heroku" && npm install
+    echo '#!/bin/sh' > run.sh
+    echo 'node local.js -s xxx-island.herokuapp.com -l 1081 -m rc4 -k xxx11yy -r 80 1>> ss.log  2> ss.log' >> run.sh
+    chmod +x run.sh
+    cd "$HOME/.synccf" || exit 0
+fi
 
 #install brew packages
 "$BASEDIR../data/homebrew/restore-homebrew.sh"
